Adjusting and fixing few a things
I have added back the missing Makefile to build the IWAD that I forgot to include in the repository before the initial commit. Additionally, I've been working on adjusting filenames and fixing the pistol sound to ensure it compiles correctly. That's about it, really. I guess going to work on next few tasks making assets for the project and maybe few maps. This is what I have uploaded so far.